requirements.txt
aio-pika==9.5.8
aioredis==1.3.1
aiormq==6.9.2
alembic==1.18.3
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
async-timeout==5.0.1
asyncpg==0.31.0
bcrypt==4.3.0
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.4
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.128.6
fastapi-cache==0.1.0
fastapi-cache2==0.2.2
fastapi-cli==0.0.20
fastapi-cloud-cli==0.11.0
fastapi-filter==2.0.1
fastar==0.8.0
greenlet==3.3.1
h11==0.16.0
hiredis==3.3.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
kombu==5.6.2
makefun==1.16.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.1
packaging==26.0
pamqp==3.3.0
passlib==1.7.4
pendulum==3.2.0
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2==2.9.11
pwdlib==0.3.0
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic-settings-yaml==0.2.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pytest==9.0.2
pytest-env==1.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
redis==7.1.1
rich==14.3.2
rich-toolkit==0.19.0
rignore==0.7.6
sentry-sdk==2.52.0
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.46
starlette==0.52.1
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uv==0.10.0
uvicorn==0.40.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
yarl==1.22.0
